create open db helper function
