+++ cmd/api/main.go
@@ -1,36 +1,29 @@
 package main
 
 import (
-	"context"
-	"database/sql"
 	"log"
 	"net/http"
 	"time"
 
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/gorilla/mux"
-	"github.com/jeanpsv/realworld-golang/internal/repository/db"
+	"github.com/jeanpsv/realworld-golang/internal/repository"
+	mysqlRepository "github.com/jeanpsv/realworld-golang/internal/repository/mysql"
 	"github.com/jeanpsv/realworld-golang/internal/rest"
 	"github.com/jeanpsv/realworld-golang/services"
 )
 
 func main() {
 
-	dbConn, err := openDB("realworld:realworld@/realworld_dev?parseTime=true")
+	dbConn, err := repository.OpenDB("mysql", "realworld:realworld@/realworld_dev?parseTime=true")
 	if err != nil {
 		log.Fatal("failed to open connection to database", err)
 	}
-
-	err = dbConn.Ping()
-	if err != nil {
-		log.Fatal("failed to ping database", err)
-	}
-
 	defer dbConn.Close()
 
 	router := mux.NewRouter()
 
-	tagRepo := db.NewTagRepository(dbConn)
+	tagRepo := mysqlRepository.NewTagRepository(dbConn)
 	tagService := services.NewTagService(tagRepo)
 	rest.NewTagHandler(router, tagService)
 
@@ -42,23 +35,4 @@ func main() {
 	}
 
 	log.Fatal(server.ListenAndServe())
-}
-
-func openDB(dsn string) (*sql.DB, error) {
-	db, err := sql.Open("mysql", dsn)
-	if err != nil {
-		return nil, err
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	err = db.PingContext(ctx)
-
-	if err != nil {
-		db.Close()
-		return nil, err
-	}
-
-	return db, nil
 }
